requirements.txt
streamlit
pandas
plotly
kaleido
python-dateutil
reportlab
openpyxl
xlrd
numpy
streamlit-cookies-manager
sqlalchemy
pymysql
mysql-connector-python
